Bump the version beyond that of Lumberjack::Application
The problem is that if the module exists in the "provides" of another
module which has a higher version number (e.g. in this case the last
version of Lumberjack::Application that provided
Lumberjack::Message::JSON was v0.0.8,) will always install that "old"
version if either requested unqualified or with (e.g.) ":ver<0.0.3+>"
until the "new" module has a higher version number.